refactor: clean up dexie code
