+++ renderer/store/schema/setting.js
@@ -1,0 +1,10 @@
+/**
+ * @class Setting
+ * Setting helper class.
+ */
+export default class Setting {
+  static SCHEMA = {
+    key: String,
+    value: String,
+  }
+}
